+++ app/src/main/AndroidManifest.xml
@@ -27,11 +27,18 @@
             </intent-filter>
         </activity>
         <activity
+            android:name=".Login"
+            android:label="@string/mi_activity_login" />
+        <activity
             android:name=".Home"
             android:label="@string/mi_activity_home" />
         <activity
-            android:name=".Login"
-            android:label="@string/mi_activity_login" />
+            android:name=".IA"
+            android:label="@string/mi_activity_ia" />
+        <activity
+            android:name=".GPS"
+            android:label="@string/mi_activity_gps" />
+
     </application>
 
 </manifest>
