+++ app/src/main/java/com/example/bachefinder/GPS.java
@@ -1,0 +1,11 @@
+package com.example.bachefinder;
+
+import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class GPS extends AppCompatActivity {
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_gps);
+    }
+}
